Initial version of check for tests & global vars
We want to ensure that any test that alters the value of a global
variable restores it to its original state.  We even have a helper class
to ensure that happens, even in the case of exceptions.

This is a first draft of a clang-tidy check intended to verify that
tests follow this rule.

(cherry picked from commit 6f1a5ffa95dccc39b96879b17b6fb63cb01b847b)